+++ libs/Engine/src/tools.cpp
@@ -3,14 +3,15 @@
 #include <utility>
 #include <vector>
 
-[[nodiscard]] vec vec::cord(int width, int height) const {  // преобразование из моих координат в сфмл
-    double aspect = static_cast<double>(width) / height;
-    return {(x + 1. * aspect) / aspect / 2. * width, (-1 * y + 1.) / 2. * height};
+[[nodiscard]] vec vec::cord(
+    int width, int height) const {  // преобразование из моих координат в сфмл
+    float aspect = static_cast<float>(width) / height;
+    return {(x + 1.f * aspect) / aspect / 2.f * width, (-1.f * y + 1.f) / 2.f * height};
 }
 
 [[nodiscard]] vec vec::anticord(int width, int height) const {  // обратное
-    double aspect = static_cast<double>(width) / height;
-    return {x * aspect * 2. / width - 1. * aspect, (-1 * y) * 2. / height + 1.};
+    float aspect = static_cast<float>(width) / height;
+    return {x * aspect * 2.f / width - 1.f * aspect, (-1.f * y) * 2.f / height + 1.f};
 }
 
 [[nodiscard]] line line::move(const vec& vector) const {
@@ -18,16 +19,16 @@
 }
 
 [[nodiscard]] vec line::norm() const {  // единичная нормаль к поверхности
-    double a = p1.y - p2.y, b = p2.x - p1.x;
+    float a = p1.y - p2.y, b = p2.x - p1.x;
     vec no(a, b);
     return no / no.length();
 }
 
-double Distance(const vec& p1, const vec& p2) {
+float Distance(const vec& p1, const vec& p2) {
     return sqrt((p2.x - p1.x) * (p2.x - p1.x) + (p2.y - p1.y) * (p2.y - p1.y));
 }
 
-double Distance(const vec& point, const line& segment) {
+float Distance(const vec& point, const line& segment) {
     if ((point - segment.p1) * (segment.p2 - segment.p1) < 0) {
         return (point - segment.p1).length();
     }
@@ -35,18 +36,18 @@ double Distance(const vec& point, const line& segment) {
         return (point - segment.p2).length();
     }
 
-    double a = segment.p1.y - segment.p2.y, b = segment.p2.x - segment.p1.x,
-           c = segment.p1.x * segment.p2.y - segment.p2.x * segment.p1.y;
+    float a = segment.p1.y - segment.p2.y, b = segment.p2.x - segment.p1.x,
+          c = segment.p1.x * segment.p2.y - segment.p2.x * segment.p1.y;
     return abs(a * point.x + b * point.y + c) / sqrt(a * a + b * b);
 }
 
-double SignedDistance(const vec& point, const line& segment) {
-    double a = segment.p1.y - segment.p2.y, b = segment.p2.x - segment.p1.x,
-           c = segment.p1.x * segment.p2.y - segment.p2.x * segment.p1.y;
+float SignedDistance(const vec& point, const line& segment) {
+    float a = segment.p1.y - segment.p2.y, b = segment.p2.x - segment.p1.x,
+          c = segment.p1.x * segment.p2.y - segment.p2.x * segment.p1.y;
     return (a * point.x + b * point.y + c) / sqrt(a * a + b * b);
 }
 
-double Distance(const line& l1, const line& l2) {
+float Distance(const line& l1, const line& l2) {
     if (IsIntersect(l1, l2)) {
         return 0;
     }
@@ -56,11 +57,11 @@ double Distance(const line& l1, const line& l2) {
 }
 
 bool IsIntersect(const line& l1, const line& l2) {
-    double a1 = SignedDistance(l1.p1, l2);
-    double b1 = SignedDistance(l1.p2, l2);
+    float a1 = SignedDistance(l1.p1, l2);
+    float b1 = SignedDistance(l1.p2, l2);
 
-    double a2 = SignedDistance(l2.p1, l1);
-    double b2 = SignedDistance(l2.p2, l1);
+    float a2 = SignedDistance(l2.p1, l1);
+    float b2 = SignedDistance(l2.p2, l1);
 
     return (a1 >= 0 && b1 <= 0 || a1 <= 0 && b1 >= 0) && (a2 >= 0 && b2 <= 0 || a2 <= 0 && b2 >= 0);
 }
@@ -92,7 +93,7 @@ void MoveableObject::sumNormalForces(Object* obj) {
         line1 = line1.move(basepoint);
         for (line line2 : obj->polygons) {
             line2 = line2.move(obj->basepoint);
-            vec Nadditional;  
+            vec Nadditional;
             if (Distance(line1, line2) <= PH_CONST_COLLISION_PRES) {
                 if (Projection(velocity, line2.norm()) * (basepoint - line2.p1) >= 0) {
                     continue;  // make obj get out of other obj
